Use counts estimated from each analysis
Previously I used the normalised counts from the analysis with all
samples and filtered to get the time points I wanted. This is a bit
misleading because counts are normalised by samples that aren’t
plotted. A bigger problem is that only genes with reasonable counts in
samples from all time points are included. This leads to situations
where a gene shows up in one of the DE tables but can’t be plotted
because it is not expressed at other time points. Now completely
different count estimates are used for each time point so the numbers
for a given sample will be slightly different when it is plotted in the
full dataset or the single week analysis