Sort out problems with the background with DESeq
There was a bug with the way I was filtering out genes with hight Cooks
distances. This should be sorted out now. The numbers aren’t identical
because of rounding errors, but the size of the Background file is
roughly the size of complete dataset less the number of features
removed through independent filtering.